+++ tests/testthat/test-weights.R
@@ -10,3 +10,13 @@ test_that("invalid 'weights'", {
   expect_error(el_mean(x, par, weights = w))
   expect_error(el_mean(x, par, weights = rep("w", length(x))))
 })
+
+test_that("weights() function", {
+  data("airquality")
+  x <- airquality$Wind
+  w <- airquality$Day
+  n <- length(x)
+  expect_identical(weights(el_mean(x, par = 10)), NULL)
+  fit <- el_mean(x, par = 10, weights = w)
+  expect_equal(sum(weights(fit)), length(x))
+})