+++ NEWS.md
@@ -6,12 +6,14 @@ melt 1.6.0
 ### NEW FEATURES
 * New package dependencies added (`graphics`, `lifecycle`, `BH`, and `dqrng`).
 
+* `elmt()` is added for multiple testing with empirical likelihood.
+
 * Generic function `weights()` is added.
 
 * `lht()` is renamed to `elt()`, which accepts additional arguments `alpha` and `calibrate`.
 
 ### MINOR IMPROVEMENTS
-* `cv` argument in `confint()` and `confreg()` defaults to NULL.
+* `cv` argument in `confint()` and `confreg()` defaults to `NULL`. If non-`NULL`, `level` is ignored.
 
 ### BUG FIXES
 * unit test error fixed.